backend/requirements.txt
fastapi
uvicorn
fastembed
pydantic
underthesea
python-dotenv
llama-index
google-generativeai
sentence-transformers
qdrant-client
requests>=2.31.0